Refactor to simpler webpack config + use generate-examples-index-webpack-plugin
