+++ package.json
@@ -12,8 +12,8 @@
   "scripts": {
     "start": "webpack-dev-server --env=dev",
     "clean": "rimraf lib",
-    "build": "tsc",
-    "examples": "webpack --env=examples",
+    "build": "tsc && yarn docs",
+    "examples": "webpack",
     "prebuild": "yarn run clean",
     "docs": "typedoc ./src",
     "clean:docs": "rimraf docs",
@@ -32,16 +32,15 @@
     "@types/node": "^10.5.2",
     "@types/webpack-env": "^1.13.5",
     "clean-webpack-plugin": "^0.1.19",
-    "copy-webpack-plugin": "^4.5.2",
     "css-loader": "^1.0.0",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "fontfaceobserver": "^2.0.13",
+    "generate-examples-index-webpack-plugin": "^1.0.1",
     "html-webpack-plugin": "^3.2.0",
     "include-all": "^4.0.3",
     "less": "^3.7.1",
     "less-loader": "^4.0.5",
     "merge-files-webpack-plugin": "^1.1.2",
-    "mkdirp": "^0.5.1",
     "node-sass": "^4.9.2",
     "postcss-import": "^11.1.0",
     "postcss-loader": "^2.1.6",
