+++ webpack.settings.js
@@ -1,4 +1,12 @@
-const absPath = require('./util/absPath');
+const path = require('path');
+
+/**
+ * @param {string} pathFromProjectRoot path specified from the root of the project (where package.json is)
+ * @return {string} absolute path
+ */
+function absPath(pathFromProjectRoot) {
+  return path.resolve(__dirname, pathFromProjectRoot);
+}
 
 module.exports = {
   paths: {
